realistic test for parameterized evm schema.
note: discovered wrongly schema'd eth_txs types
