+++ core/store/migrate/plugins/relayer/evm/0001_create_schema.tmpl.sql
@@ -1,0 +1,4 @@
+-- +goose Up
+CREATE SCHEMA IF NOT EXISTS {{.Schema}};
+-- +goose Down
+DROP SCHEMA IF EXISTS {{.Schema}};
