+++ core/store/migrate/plugins/relayer/evm/0002_initial.go
@@ -7,21 +7,11 @@ import (
 	_ "embed"
 	"fmt"
 	"io"
-	"text/template"
 
 	"github.com/pressly/goose/v3"
 )
 
-type Cfg struct {
-	Schema  string
-	ChainID int
-}
-
-func RegisterSchemaMigration(val Cfg) error {
-	return Register(val)
-}
-
-// go:embed initUp.tmpl.sql
+//go:embed initUp.tmpl.sql
 var upTmpl string
 
 func resolveUp(out io.Writer, val Cfg) error {
@@ -31,24 +21,11 @@ func resolveUp(out io.Writer, val Cfg) error {
 	return resolve(out, upTmpl, val)
 }
 
-// go:embed  initDown.tmpl.sql
+//go:embed  initDown.tmpl.sql
 var downTmpl string
 
 func resolveDown(out io.Writer, val Cfg) error {
 	return resolve(out, downTmpl, val)
-}
-
-func resolve(out io.Writer, in string, val Cfg) error {
-	id := fmt.Sprintf("init_%s_%d", val.Schema, val.ChainID)
-	tmpl, err := template.New(id).Parse(in)
-	if err != nil {
-		return fmt.Errorf("failed to parse template %s: %w", in, err)
-	}
-	err = tmpl.Execute(out, val)
-	if err != nil {
-		return fmt.Errorf("failed to execute template %s: %w", in, err)
-	}
-	return nil
 }
 
 // Register registers the migration with goose
@@ -59,8 +36,6 @@ func Register(val Cfg) error {
 		return fmt.Errorf("failed to resolve up sql: %w", err)
 	}
 	upFunc := func(ctx context.Context, tx *sql.Tx) error {
-		fmt.Printf("Executing up sql: %s\n", upSQL.String())
-		panic(fmt.Sprintf("Executing up sql: %s\n", upSQL.String()))
 		_, err := tx.ExecContext(ctx, upSQL.String())
 		return err
 	}
