+++ core/store/migrate/plugins/relayer/evm/initDown.tmpl.sql
@@ -1,3 +1,4 @@
+/*
 DROP TABLE {{ .Schema }}.receipts
 DROP TABLE {{ .Schema }}.tx_attempts
 DROP TABLE {{ .Schema }}.upkeep_states
@@ -7,4 +8,11 @@ DROP TABLE {{ .Schema }}.log_poller_filters
 DROP TABLE {{ .Schema }}.log_poller_blocks
 DROP TABLE {{ .Schema }}.key_states;
 DROP TABLE {{ .Schema }}.heads;
+*/
+
+-- Copy data from old table to new table
+INSERT INTO evm.forwarders (address, created_at, updated_at, evm_chain_id)
+SELECT address, created_at, updated_at, '{{ .ChainID }}' 
+FROM {{ .Schema }}.forwarders;
+
 DROP TABLE {{ .Schema }}.forwarders;
